refactor(application): extract payment confirmation use case from webhook service
